moving lists into a fragment
