+++ app/src/main/java/com/oliverbud/android/imagelist/MainActivity.java
@@ -30,6 +30,7 @@ import com.oliverbud.android.imagelist.Application.App;
 import com.oliverbud.android.imagelist.EventBus.AddItemsEvent;
 import com.oliverbud.android.imagelist.EventBus.GenericEvent;
 import com.oliverbud.android.imagelist.EventBus.NavItemSelectedEvent;
+import com.oliverbud.android.imagelist.EventBus.SearchEvent;
 
 import java.lang.reflect.Array;
 import java.util.ArrayList;
@@ -41,13 +42,13 @@ import butterknife.InjectView;
 import butterknife.Optional;
 import dagger.ObjectGraph;
 import de.greenrobot.event.EventBus;
+import de.greenrobot.event.util.ErrorDialogManager;
 import icepick.Icepick;
 
 
-public class MainActivity extends AppCompatActivity implements ImageListView{
+public class MainActivity extends AppCompatActivity{
 
     @InjectView(R.id.searchInput)Toolbar searchInput;
-    @InjectView(R.id.listPager) ViewPager listPager;
     @Optional @InjectView(R.id.drawerLayout) DrawerLayout drawerLayout;
     @Optional @InjectView(R.id.linearLayout) LinearLayout linearLayout;
     @InjectView(R.id.tabLayout) TabLayout tabLayout;
@@ -55,7 +56,6 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
     @InjectView(R.id.spinner)ProgressBar spinner;
 
 
-    listPagerAdapter myPagerAdapter;
 
     String currentSearch = null;
     ArrayList<String> searchStrings;
@@ -75,14 +75,16 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
         Icepick.restoreInstanceState(this, savedInstanceState);
 
         if (isTablet(this)){
-            setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE);
+//            setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE);
             setContentView(R.layout.activity_main_tablet_landscape);
             ButterKnife.inject(this);
             setSupportActionBar(searchInput);
+            tabLayout.setupWithViewPager(((ListsDisplayFragment)getSupportFragmentManager().findFragmentByTag("listsFragment")).listsViewPager);
+
 
         }
         else {
-            setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
+//            setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
             setContentView(R.layout.activity_main_phone_protrait);
             ButterKnife.inject(this);
             setSupportActionBar(searchInput);
@@ -91,35 +93,23 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
             getSupportActionBar().setHomeButtonEnabled(true);
             abdt.syncState();
-
+            tabLayout.setupWithViewPager(((ListsDisplayFragment)getSupportFragmentManager().findFragmentByTag("listsFragment")).listsViewPager);
         }
 
         spinner.setIndeterminate(true);
         spinner.setVisibility(View.GONE);
 
-
-        activityGraph = ((App) this.getApplication()).createScopedGraph(new PresenterModule(MainActivity.this));
-        activityGraph.inject(this);
-
-
-        searchStrings = new ArrayList<String>();
+        searchStrings = new ArrayList();
 
 
 
         if (savedInstanceState != null){
             currentSearch = savedInstanceState.getString("currentSearchString");
             searchStrings = savedInstanceState.getStringArrayList("searchStrings");
-            presenter.list = savedInstanceState.getParcelableArrayList("presenterList");
-            presenter.page = savedInstanceState.getInt("page");
+
             EventBus.getDefault().post(new AddItemsEvent(searchStrings));
 
         }
-
-
-
-
-
-        initializePaging();
 
         handleIntent(getIntent());
 
@@ -131,14 +121,7 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
                 >= Configuration.SCREENLAYOUT_SIZE_LARGE;
     }
 
-    public void initializePaging(){
-        Log.d("itemListApp", "initializePaging");
 
-        myPagerAdapter = new listPagerAdapter();
-        listPager.setAdapter(myPagerAdapter);
-        listPager.setOffscreenPageLimit(2);
-        tabLayout.setupWithViewPager(listPager);
-    }
 
     @Override
     public void onBackPressed() {
@@ -162,10 +145,7 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
             String query = intent.getStringExtra(SearchManager.QUERY);
             Log.d("itemListApp", "handling intent: " + query);
 
-            for (int i = 0; i < 3; i ++){
-                myPagerAdapter.setLoadTrue(i);
-            }
-            presenter.searchFor(query);
+
 
             MenuItem menuItem = this.menu.findItem(R.id.search);
             menuItem.collapseActionView();
@@ -174,6 +154,7 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
             ArrayList<String> addItems = new ArrayList<String>();
             addItems.add(query);
             EventBus.getDefault().post(new AddItemsEvent(addItems));
+            EventBus.getDefault().post(new SearchEvent(query));
 
 
             currentSearch = query;
@@ -200,8 +181,7 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
 
         outState.putString("currentSearchString", currentSearch);
         outState.putStringArrayList("searchStrings", searchStrings);
-        outState.putParcelableArrayList("presenterList", presenter.list);
-        outState.putInt("page", presenter.page);
+
         super.onSaveInstanceState(outState);
         Icepick.saveInstanceState(this, outState);
 
@@ -221,21 +201,6 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
         EventBus.getDefault().register(this);
     }
 
-    public void onEvent(GenericEvent event) {
-        if (event.event != "SNACKS") {
-            View.OnClickListener myOnClickListener = new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    drawerLayout.openDrawer(GravityCompat.START);
-                }
-            };
-
-            Snackbar
-                    .make(coordinatorLayout, "HULLO", Snackbar.LENGTH_LONG)
-                    .setAction("CLICK", myOnClickListener)
-                    .show();
-        }
-    }
 
     public void onEvent(NavItemSelectedEvent event) {
 
@@ -246,13 +211,6 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
             drawerLayout.closeDrawer(GravityCompat.START);
         }
         if (!searchParam.equals(currentSearch)) {
-            for (int i = 0; i < 3; i ++){
-                myPagerAdapter.setLoadTrue(i);
-            }
-            presenter.searchFor(searchParam);
-
-            EventBus.getDefault().post(new GenericEvent("SNACKS"));
-
 
             searchInput.setTitle(searchParam);
             currentSearch = searchParam;
@@ -299,125 +257,8 @@ public class MainActivity extends AppCompatActivity implements ImageListView{
         return super.onOptionsItemSelected(item);
     }
 
-    @Override
-    public void setItems(ArrayList<ImageDataItem> listData) {
-        listPager.setVisibility(View.VISIBLE);
-        spinner.setVisibility(View.GONE);
-        for (int i = 0; i < 3; i ++){
 
 
-            myPagerAdapter.setListAdapterForPosition(new ImageListAdapter(listData), i);
-        }
 
-    }
-
-    @Override
-    public void addItems(ArrayList<ImageDataItem> listData) {
-
-        for (int i = 0; i < 3; i ++){
-            myPagerAdapter.updateAdapter(i);
-        }
-
-    }
-
-    @Override
-    public void displayLoading() {
-        Log.d("itemListApp", "displayLoading Activity");
-        listPager.setVisibility(View.GONE);
-        spinner.setVisibility(View.VISIBLE);
-
-    }
-
-    @Override
-    public void displayError() {
-        for (int i = 0; i < 3; i ++) {
-            myPagerAdapter.setLoadFalse(i);
-        }
-        Log.d("itemListApp", "displayError Activity");
-    }
-
-
-    public class listPagerAdapter extends PagerAdapter{
-
-        SmartListView[] listViews = new SmartListView[3];
-        ImageListAdapter[] listViewAdapters = new ImageListAdapter[3];
-
-
-        public void setListAdapterForPosition(ImageListAdapter ila, int position){
-            listViews[position].setAdapter(ila);
-            listViewAdapters[position] = ila;
-        }
-
-        public void updateAdapter(int i ) {
-            if (listViewAdapters[i] != null) {
-                listViewAdapters[i].notifyDataSetChanged();
-
-            }
-        }
-
-        public void setLoadFalse(int i){
-            listViews[i].getScrollListener().stopLoading();
-        }
-
-        public void setLoadTrue(int i){
-            listViews[i].getScrollListener().tryLoading();
-        }
-
-        @Override
-        public boolean isViewFromObject(View view, Object object) {
-            return view == object;
-        }
-
-        @Override
-        public int getCount() {
-            return 3;
-        }
-
-        @Override
-        public CharSequence getPageTitle(int position) {
-            switch (position){
-                case 0:
-                    return "bacon";
-                case 1:
-                    return "salad";
-                case 2:
-                    return "spoons";
-
-            }
-            return "default";
-        }
-
-        @Override
-        public Object instantiateItem(ViewGroup container, int position) {
-            SmartListView lv = new SmartListView(MainActivity.this);
-
-            listViews[position] = lv;
-
-            lv.setScrollListener(new EndlessScrollListener(lv.getLayoutManager()) {
-                @Override
-                public void onLoadMore(int current_page) {
-                    MainActivity.this.presenter.loadMore(currentSearch);
-
-                }
-            });
-
-            if (presenter.list != null){
-                lv.setAdapter(new ImageListAdapter(presenter.list));
-            }
-
-
-            ViewGroup.LayoutParams layoutParams = new ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT);
-            container.addView(lv, layoutParams);
-            return lv;
-        }
-
-        @Override
-        public void destroyItem(ViewGroup container, int position, Object object) {
-            listViews[position] = null;
-            listViewAdapters[position] = null;
-
-            super.destroyItem(container, position, object);
-        }
-    }
 
 }
