+++ app/src/main/res/layout/activity_main_tablet_landscape.xml
@@ -22,6 +22,7 @@
 
 
 
+
         <android.support.design.widget.AppBarLayout
             android:id="@+id/appBar"
             android:background="@color/colorPrimary"
@@ -43,11 +44,11 @@
         </android.support.design.widget.AppBarLayout>
 
 
-
-    <android.support.v4.view.ViewPager
-            android:layout_width="match_parent"
+        <fragment
+            android:layout_width="wrap_content"
             android:layout_height="match_parent"
-            android:id="@+id/listPager"
+            android:name="com.oliverbud.android.imagelist.ListsDisplayFragment"
+            android:tag="listsFragment"
             app:layout_behavior="@string/appbar_scrolling_view_behavior"/>
 
         <ProgressBar
@@ -55,17 +56,8 @@
             android:layout_height="60dp"
             android:id="@+id/spinner"/>
 
-
-
-
     </android.support.design.widget.CoordinatorLayout>
 
-    <!--<android.support.design.widget.NavigationView-->
-    <!--android:layout_width="wrap_content"-->
-    <!--android:layout_height="match_parent"-->
-    <!--android:layout_gravity="start"-->
-    <!--android:id="@+id/navigation"-->
-    <!--app:menu="@menu/drawer"/>-->
 </LinearLayout>
